Tests/LibWeb/Text/expected/css/shadow-media-query-invalidation-stays-scoped.txt

Target color: rgb(0, 128, 0)
PASS: shadow media query invalidation stayed scoped (12 invalidations)
